app/requirements.txt
async-timeout==4.0.2
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-WTF==0.15.1
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
redis==4.5.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
